% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flsgen_available_targets.R
\docType{data}
\name{CLASS_LEVEL_TARGETS}
\alias{CLASS_LEVEL_TARGETS}
\title{Vector of available class targets}
\format{
An object of class \code{character} of length 16.
}
\usage{
CLASS_LEVEL_TARGETS
}
\description{
Vector of available class targets
}
\keyword{datasets}
