% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flsgen_create_class_structure.R
\name{flsgen_create_class_structure}
\alias{flsgen_create_class_structure}
\title{Creates a predefined landscape class structure that can be converted as
JSON input for flsgen generate.}
\usage{
flsgen_create_class_structure(class_name, patch_areas)
}
\arguments{
\item{class_name}{Name of the class}

\item{patch_areas}{vector of patch areas}
}
\value{
A landscape class structure
}
\description{
Creates a predefined landscape class structure that can be
 converted as JSON input for flsgen generate.
}
\examples{
  \dontrun{
    cls_1 <- flsgen_class_structure("class 1", c(10, 100, 1000))
  }

}
